{nixos/services,shared/ai-grammar-fix}: add AI grammar and spell checker
The idea behind this Bash script is to emulate the Raycast
(a macOS Spotlight alternative) feature for fixing grammar and spelling
by simply selecting text and invoking the command. In our case,
we are using the clipboard content as the input source

The way it works is:

- Grabs the last clipboard content and saves it to a temporary .TXT
  file
- Uses the Phi-3 with a settings to correct the text
- Saves the corrected text to another .TXT file and compares it
  with the original
- Sends a notification if changes are made (or not) and copies the
  corrected text back to the clipboard

Note: Frankly, even in this super niche scenario, AI still isn't that good.
So this is more geared towards serving as a "skeleton," per se,
to more quickly and easily spot any obviously or less obvious spelling
or grammar mistakes you might have missed, and for the most part,
it still requires manual editing to really take advantage of
the black magic output the AI spews back at you

PS: By AI, I *largely* mean LLMs (Large Language Models)