hosts/darwin/homebrew: `firefox` -> `zen-browser`
Firefox is easily my favorite browser, and there are many things I can
only do on Firefox but not on Chromium browsers. BUT, sadly, Firefox is
becoming very bloated, and it's real pain to use on macOS, especially on
machines with 8 GiB of RAM

Zen Browser, on the other hand, is a Firefox fork that's just perfect
for me. It's a very lightweight version of Firefox with built-in
Tree Style Tabs and runs macOS. I wish it had existed sooner