modules/hm/guis/vscode/extensions/default: switch to nixpkgs `copilot{-,-chat}`
`nix-vscode-extensions` has a bug where extensions disappear from the
JSON in the repo, making them unusable. Usually, this isn't a big issue
since the extensions return in 1-3 days. BUT the `copilot-chat`
extension hasn't returned and doesn't seem like it will. So, we're
switching back to the nixpkgs `copilot-chat`

We're switching to the `copilot` extension from nixpkgs to avoid any
version mismatches