hosts/darwin/{default,homebrew}: drop homebrew module
I finally got around to this, though I should've done it earlier. The
catalyst was helping a friend set up Nix on his Darwin machine from
scratch. While recommending software I use, it dawned on me that I
*might* not need Homebrew anymore. I could use `nixpkgs` for all my
package needs, and manage stuff that `nixpkgs` doesn't have imperatively
instead of dealing with Homebrew's massive overhead.

I'll be managing Telegram, Rider, and Warp imperatively. I really need
their latest version available as soon as possible and don't want to
rely on nixpkgs bureaucracy. Plus, Telegram on Swift isn't available on
`nixpkgs`, so there's that...

The reality is that macOS is inherently impure as a system, and what I'm
doing to it is admittedly quite hacky. But honesetly I'm *QUITE* happy
to trade some purity for the massive convenience of eliminating
Homebrew.

Another nice convenience I will be getting out of this is that I won't
be prompted to enter my password twice, and I won't have to wait an
extra long time for Homebrew.