treewide: add `starship`, drop `powerlevel10k`
HAPPY NEW YEAR!!!

Well, this is my first commit of 2025, and I think it's a pretty good
commit to start the year on! It's part of a larger goal I have, which is
to basically TUIfiy, Rustify, Zigify, and modernize all of the CLI tools
I use.

The main reason is that right now we're in a very sweet spot where Linux
has gained so many new users, many of whom are great developers and have
created amazing tools with these languages!

Specifically for Rust, for example, it has amazing error handling at the
compiler and language level, which seemingly all Rust programs get to
inherit! Another great thing is that stuff written in Rust and Zig
usually transcends any OS lock-in and can provide the same interface on
all three systems. Take Starship for our case - Starship is able to run
in all shells on all systems. How amazing is that?? I get to have the
same prompt on `zsh`, `bash`, `fish`, `nushell`, and so on AND on Linux,
macOS, **AND WINDOWS?!**

It just doesn't get better - it's the obvious choice over using
powerlevel10k, which is locked to zsh. Which is fine since I mainly use
that, **BUT** I will most likely be moving over to NuShell (which is
also written in Rust) and again is transferable to other systems!

Honestly, before I didn't really understand why people "jerk off" Rust
so much - what was so good about it anyway? It's just another language!
Well, the last part is still true, **BUT** seemingly all the re-written
tools or just new tools created in it get to have amazing error handling
AND have universal interfaces across multiple platforms. Just simply
amazing!

I think in 2025, this is **HOW** I want to "treat myself," so to speak.
I'm honestly only going to use GNU implementations for scripts that are
meant to be consumed by other people and are meant to be general
purpose, BUT FOR ME personally? I'm going to be looking to replace ALL
GNU implementations. I want to move on - I want my terminal to live in
2025 too!

There's also the performance argument, too. Funnily enough, the Rust
implementation, which is technically written in a higher-level language,
is **MORE** performant and has better-generated Assembly code.

Well, that's it! Thank you for reading all of this, and again, Happy New
Year!!!!