treewide: add `helix`
A post-modern replacement for Neovim, basically what if we remade
Neovim in Rust and implemented the most used plugins as native features,
treating them as first-class citizens

Honestly, it's a pretty nice experience compared to NeoVim. It loads up
really fast, and I don't have to worry about LSPs, DAPs, formatters, and
so on.

It's also part of my ongoing rustification of my setup too!