shared/scripts/ai-grammar-fix.sh: drop
So the backstory of this script is that it was supposed to be a script
that I would trigger with some keybind to run a local model with
`ollama` to fix my grammar. That never really panned out, and I kind of
gave up on the idea as well. If I ever come back to this, I will most
likely use an API since the prices now are so cheap, and I think also
Google's API pricing is very generous, as they give you a ton of free
calls per day or month