{hosts/{darwin,nixos}/home,hosts/darwin/zsh,shared/aliases}: refactor
Instead of using `pkgs` to get `isLinux`, we're using
`osConfig.nixpkgs.hostPlatform.isLinux` to not be dependent on pkgs

Removed `nixCfgPath`. I realized that it's much more standard to
actually make `/etc/nixos` a symlink to wherever I'm actually keeping my
config; that way, it's universal and doesn't require conditionals

Instead of using `home.shellAliases`, I'm now using
`programs.{bash,zsh}` directly. The reason is that previously,
`home.shellAliases` only included Bash and Zsh, but it does not include
NuShell, which means many of the aliases I use for Bash and Zsh won't
work since NuShell isn't POSIX compliant

I also got rid of `mergeAttrs`, so I can't remember anymore why I wanted
to do this, but basically, it was to escape any alias that had a space.
Looking back on it, it was kind of pointless. Most aliases don't need to
be escaped, and I can just use `lib.escapeShellArg` if I wanted to...

I also got rid of the aliases that call Rust utils instead of GNU ones.
At first, I thought this would be a good idea, but often what happened
was that whenever I needed GNU variants, I would always have to `\` on
{Bash, Zsh} and `^` on NuShell. It's honestly a better idea to remember
them by their real names, and I actually even gaslighted myself into
thinking `dust` was actually `du` because that's how I was always
calling it. So, not good...

And finally, I'm using `pkgs.callPackage` instead of the raw `import`,
which automatically supplies `pkgs` and `lib` for me