{flake.*,modules/hm/cli/git}: add `jj-vsc` input, `jj template` config
Since `jj` is an ever-evolving project, it's beneficial to stay as
up-to-date as possible. Therefore, the package is now configured to use
Tip-of-Tree (ToT)

One key reason for this update is the need for `jj`'s new "templates"
feature. This allows replicating a Git backporting workflow where `git
cherry-pick -x` is used. The `-x` flag adds a message like `(cherry
picked from commit <hash>)` to the commit body. This addition is _very_
important for tracking where a commit originated. The new `jj` template
configuration achieves this same result