{flake.*,modules/common}: add lix module
A while ago, I tried out Lix, but I wasn't really satisfied with it.
_BUT_, after reading the last two release notes (2.92) & (2.93), I
feel like Lix finally has some nice stuff to offer, especially for
macOS/Darwin. So, I'm going to give it a try again and see how it goes

For macOS/Darwin specfically these are the stuff that drew me in:

- `auto-optimise-store` corruption fixed on macOS
- Ctrl-C on macOS no longer leaves nix-daemon processes lingering
- `trusted-users` and `allowed-users` group checks fixed on macOS
- `nix-daemon` fetches client PID on macOS