{flake.*,hosts/{darwin,nixos}/home}: bump, `nixos-unstable-small` -> `25.05`
Welp, NixOS Release 25.05 is almost out. ZHF isn't fully done yet, but
as of this writing, there are only 6 days left until it's finished, and
then a few more days until the final release

_For all practical purposes, though, it's already here_. `25.05` is
likely the version I'll stick with for stability

Until now, the stable version always felt like it lagged significantly
behind unstable, but that's no longer the case. This is actually a good
thing, as the NixOS module system has grown a lot. I feel like `25.05`
will be a great release to "settle down" on for its duration.

Hopefully, I won't need to switch back to unstable, since there are real
benefits to sticking with a stable release instead of constantly running
unstable and having to update your modules manually every now and then