feat(hm/terminal): add zellij multiplexer integration
See, until now I was relying on Ghostty as my poor man's tab/pane
manager and I honestly liked it, but even from the get-go I knew that
even compared to tmux what _I had was not that great really_ - it's kind
of like a knock-off but I honestly stuck with it because it just worked
and I got comfortable with it.

_BUT_ that's what I _HATE_ in a way - I hate being comfortable
because when I'm comfortable I'm not really learning new stuff, I'm
not experiencing new stuff. It feels nice to be comfortable, it feels
reassuring and you're fast, but I don't wanna be comfortable for too
long. I don't wanna be forever uncomfortable but I want to have a
"limit" of sorts and move on to find something newer and better, and
here we are today!

So today I'm switching to Zellij - it's not as comfortable as my Ghostty
setup yet but I managed to actually bring it pretty close. I've added a
new zellij.nix module that mirrors my familiar Ghostty keybinding style
(Super+key on Darwin), updated Ghostty to launch Nu shell through Zellij
automatically, and completely ripped out Ghostty's split functionality
since Zellij handles it so much better.

One nice benefit and huge consideration is that with Zellij I'm
unchaining myself from Ghostty. See, while using Ghostty as my pane/tab
manager was nice and comfy, it meant I was stuck with using Ghostty, and
switching to any other terminal - be it even temporarily - I felt like I
was limited, that my workflow was downgraded, and I hated that feeling!

I could feel how I was becoming reliant on Ghostty, and Ghostty hasn't
even been around for that long yet, so I'm kind of locking myself into
a project that doesn't yet have established history. Not to say that
it won't - and Ghostty certainly looks like it will have established
history - nevertheless I don't wanna "bet" on stuff like that tbh.

And honestly, even outside of that, again I wanna make it easier to
migrate if I have to or want to, and also Zellij can offer me more in
the pane/tab managing department as that's one of its core features
after all! Meanwhile in Ghostty it feels like more of an afterthought
and nice extra really...